+++ src/PrimaryButton.js
@@ -2,22 +2,6 @@ import React from 'react';
 import { node } from 'prop-types';
 import Button from './Button';
 
-export default function PrimaryButton(props) {
-    const {
-        children,
-        ...rest
-    } = props;
+const PrimaryButton = props => <Button buttonType="primary" {...props} />;
 
-    return (
-        <Button
-            {...rest}
-            buttonType="primary"
-        >
-            {children}
-        </Button>
-    );
-}
-
-PrimaryButton.propTypes = {
-    children: node,
-};
+export default PrimaryButton;